extract label utils for filter clause
