+++ src/client/components/filter-tile/time-shift-label.ts
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2017-2022 Allegro.pl
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { isTimeAttribute } from "../../../common/models/data-cube/data-cube";
+import { Dimension } from "../../../common/models/dimension/dimension";
+import { Essence } from "../../../common/models/essence/essence";
+
+export default function timeShiftLabel(dimension: Dimension, essence: Essence) {
+  if (!isTimeAttribute(essence.dataCube, dimension.expression)) return null;
+  if (!essence.hasComparison()) return null;
+  return `(Shift: ${essence.timeShift.getDescription(true)})`;
+}
